refactor: change the namespace to customers and standarize action name pattern
